Remove unused helpers
