+++ src/Files.Uwp/Helpers/BitmapHelper.cs
@@ -1,4 +1,5 @@
 ﻿using Files.Filesystem;
+using Files.Filesystem.StorageItems;
 using System;
 using System.IO;
 using System.Threading.Tasks;
@@ -67,6 +68,48 @@ namespace Files.Helpers
             fileStream.Size = 0;
 
             await RandomAccessStream.CopyAsync(memStream, fileStream);
+        }
+
+        /// <summary>
+        /// This function encodes a software bitmap with the specified encoder and saves it to a file
+        /// </summary>
+        /// <param name="softwareBitmap"></param>
+        /// <param name="outputFile"></param>
+        /// <param name="encoderId">The guid of the image encoder type</param>
+        /// <returns></returns>
+        public static async Task SaveSoftwareBitmapToFile(SoftwareBitmap softwareBitmap, BaseStorageFile outputFile, Guid encoderId)
+        {
+            using IRandomAccessStream stream = await outputFile.OpenAsync(FileAccessMode.ReadWrite);
+            // Create an encoder with the desired format
+            BitmapEncoder encoder = await BitmapEncoder.CreateAsync(encoderId, stream);
+
+            // Set the software bitmap
+            encoder.SetSoftwareBitmap(softwareBitmap);
+
+            try
+            {
+                await encoder.FlushAsync();
+            }
+            catch (Exception err)
+            {
+                const int WINCODEC_ERR_UNSUPPORTEDOPERATION = unchecked((int)0x88982F81);
+                switch (err.HResult)
+                {
+                    case WINCODEC_ERR_UNSUPPORTEDOPERATION:
+                        // If the encoder does not support writing a thumbnail, then try again
+                        // but disable thumbnail generation.
+                        encoder.IsThumbnailGenerated = false;
+                        break;
+
+                    default:
+                        throw;
+                }
+            }
+
+            if (encoder.IsThumbnailGenerated == false)
+            {
+                await encoder.FlushAsync();
+            }
         }
     }
 }
